+++ CHANGELOG.md
@@ -11,6 +11,7 @@ to [Semantic Versioning][semver].
 
 - Created a top level 'ibn' command as the starting point for this program
 - Output collected information from a joining musician back to stdout
+- Included a global '--date' flag to set the event date that defaults to today
 
 ### Fixed